# server/Startup/openapi.json
{
  "x-generator": "NSwag v14.2.0.0 (NJsonSchema v11.1.0.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "Your API",
    "version": "v1"
  },
  "paths": {
    "/api/auth/Login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Login",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/Register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Register",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRequestDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/Secured": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Secured",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/AdminWantsToChangePreferencesForDevice": {
      "post": {
        "tags": [
          "Device"
        ],
        "operationId": "Device_AdminWantsToChangePreferencesForDevice",
        "requestBody": {
          "x-name": "dto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminWantsToChangePreferencesForDeviceDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/AdminWantsToClearData": {
      "get": {
        "tags": [
          "Device"
        ],
        "operationId": "Device_AdminWantsToClearData",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Devicelog"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Product": {
      "post": {
        "tags": [
          "Product"
        ],
        "operationId": "Product_CreateProduct",
        "requestBody": {
          "x-name": "createProductDto",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "Product_GetAllProduct",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/Product/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "operationId": "Product_GetProductById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "AuthResponseDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "jwt"
        ],
        "properties": {
          "jwt": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "AuthRequestDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 3
          },
          "password": {
            "type": "string",
            "minLength": 4
          }
        }
      },
      "AdminWantsToChangePreferencesForDeviceDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "deviceId",
          "intervalMilliseconds",
          "unit"
        ],
        "properties": {
          "deviceId": {
            "type": "string",
            "minLength": 1
          },
          "intervalMilliseconds": {
            "type": "integer",
            "format": "int32"
          },
          "unit": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "Devicelog": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string"
          },
          "deviceid": {
            "type": "string"
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "unit": {
            "type": "string"
          }
        }
      },
      "Product": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "decimal"
          }
        }
      },
      "CreateProductDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "price": {
            "type": "number",
            "format": "decimal"
          }
        }
      },
      "ProductDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          }
        }
      },
      "StringConstants": {
        "type": "string",
        "description": "Available eventType constants"
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Enter your JWT token in the format: Bearer {your_token}",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}